refactor: left all role related data inside roles-permissions feature
